[N0_TICKET] - refactoring, removing
