--- src/main/java/com/demo/diet/configuration/RestConfiguration.java
@@ -1,9 +1,0 @@
-package com.demo.diet.configuration;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class RestConfiguration  {
-
-}
-
