+++ src/main/java/com/demo/diet/model/ApiSubError.java
@@ -1,5 +1,2 @@
 package com.demo.diet.model;
 
-public abstract class ApiSubError {
-
-}
