--- src/main/java/com/demo/diet/model/Meal.java
@@ -1,9 +1,0 @@
-package com.demo.diet.model;
-
-import lombok.Data;
-
-@Data
-public class Meal {
-    private String name;
-    private long id;
-}
